backend/requirements.txt
icrawler
wikipedia
ollama
numpy
Pillow
